requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
pytesseract==0.3.10
Pillow==10.1.0
pika==1.3.2
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.1
httpcore==0.18.0
alembic==1.12.1
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-env==1.1.3
pytest-xdist==3.5.0
pytest-timeout==2.2.0
factory-boy==3.3.0
faker==20.1.0